# hardware_store/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для магазина бытовой техники",
        "title": "Hardware Store API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth": {
            "post": {
                "description": "Проверяет учетные данные и возвращает пару токенов (Access и Refresh).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные (email и пароль)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход. Токены возвращены.",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный email или пароль",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "description": "Обрабатывает callback от Google, обменивает code на JWT-токены.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Завершение авторизации через Google",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State parameter (must match)",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT tokens",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Missing code or state",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid code or state",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/oauth/google": {
            "get": {
                "description": "Возвращает URL для редиректа пользователя на сервер авторизации Google.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Начало авторизации через Google",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CSRF state parameter",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL для редиректа на Google",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.GoogleAuthURLResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "Создает новую учетную запись пользователя в системе. Возвращает ID созданного пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса или данные не прошли валидацию",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким email уже существует",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/clients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает список всех клиентов в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Получить список всех клиентов",
                "responses": {
                    "200": {
                        "description": "Список клиентов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hardware_store_internal_web_dto.ClientResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Создаёт нового клиента вместе с адресом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Создание клиента",
                "parameters": [
                    {
                        "description": "Данные клиента",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Клиент успешно создан",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает список всех продуктов в системе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить список продуктов",
                "responses": {
                    "200": {
                        "description": "Список продуктов успешно получен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении списка",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Создаёт новый продукт в системе на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Создать новый продукт",
                "parameters": [
                    {
                        "description": "Данные продукта для создания",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Продукт успешно создан",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибки валидации полей или некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при сохранении продукта",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает список всех поставщиков в системе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Получить список поставщиков",
                "responses": {
                    "200": {
                        "description": "Список поставщиков успешно получен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении списка",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Создаёт нового поставщика вместе с его адресом в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Создать нового поставщика",
                "parameters": [
                    {
                        "description": "Данные поставщика для создания",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Поставщик успешно создан",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибки валидации полей или некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при сохранении поставщика",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "hardware_store_internal_web_dto.AddressRequest": {
            "description": "Запрос на создание нового адреса с информацией о стране, городе и улице",
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Москва"
                },
                "country": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Россия"
                },
                "street": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Технопарк, 15"
                }
            }
        },
...
        "hardware_store_internal_web_dto.ClientRequest": {
            "description": "Запрос на создание нового товара с категорией, ценой и информацией о поставщике",
            "type": "object",
            "required": [
                "name",
                "surname"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/hardware_store_internal_web_dto.AddressRequest"
                },
                "birthday": {
                    "type": "string",
                    "example": "1999-01-01"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Иван"
                },
                "surname": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Иванов"
                }
            }
        },
        "hardware_store_internal_web_dto.ClientResponse": {
            "description": "Данные клиента включая дату регистрации и ссылку на адрес",
            "type": "object",
            "properties": {
                "address_uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "birthday": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "hardware_store_internal_web_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation error"
                }
            }
        },
        "hardware_store_internal_web_dto.GoogleAuthURLResponse": {
            "description": "Содержит URL, на который нужно перенаправить пользователя для авторизации через Google",
            "type": "object",
            "properties": {
                "auth_url": {
                    "type": "string"
                }
            }
        },
...
        "hardware_store_internal_web_dto.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "hardware_store_internal_web_dto.LoginRequest": {
            "description": "Учетные данные пользователя для входа в систему",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "SecurePassword123!"
                }
            }
        },
        "hardware_store_internal_web_dto.LoginResponse": {
            "description": "Пары токенов (Access и Refresh) для аутентификации последующих запросов",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refresh_token": {
                    "type": "string",
                    "example": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
                }
            }
        },
...
        "hardware_store_internal_web_dto.ProductRequest": {
            "description": "Запрос на обновление количества товара на складе",
            "type": "object",
            "required": [
                "available_stock",
                "category_id",
                "name",
                "price",
                "supplier_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 15
                },
                "category_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Холодильник Samsung RB38A7861B1"
                },
                "price": {
                    "type": "number",
                    "example": 75990
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "hardware_store_internal_web_dto.ProductResponse": {
            "description": "Полная информация о товаре включая цену, остатки и информацию о поставщике",
            "type": "object",
            "required": [
                "available_stock",
                "last_update_date",
                "name",
                "price",
                "product_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "category": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "image": {
                    "type": "string",
                    "example": "b1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
                },
                "last_update_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string",
                    "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
                },
                "supplier": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "hardware_store_internal_web_dto.RegisterRequest": {
            "description": "Данные для создания новой учетной записи, включая email, пароль и контактную информацию",
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone_number",
                "surname"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Alice"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "SecurePassword123!"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79990000000"
                },
                "surname": {
                    "type": "string",
                    "example": "Wonderland"
                }
            }
        },
        "hardware_store_internal_web_dto.RegisterResponse": {
            "description": "Содержит уникальный идентификатор созданного пользователя",
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "hardware_store_internal_web_dto.SupplierRequest": {
            "description": "Запрос на создание нового поставщика с контактной информацией и адресом",
            "type": "object",
            "required": [
                "name",
                "phone_number"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/hardware_store_internal_web_dto.AddressRequest"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "ООО 'ТехноСнаб'"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79561234567"
                }
            }
        },
        "hardware_store_internal_web_dto.SupplierResponse": {
            "description": "Данные поставщика включая контактную информацию и ссылку на адрес",
            "type": "object",
            "properties": {
                "address_uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "ООО 'ТехноСнаб'"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79561234567"
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "hardware_store_internal_web_dto.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid uuid format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: \"Bearer \u003cваш_токен\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
                "email": "Просмотр вашего email",
                "profile": "Просмотр вашего профиля"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "OAuth2": []
        }
    ]
}